sc - add mongo properties and rename packages
