+++ javascript/.env.production.SAMPLE
@@ -1,0 +1,3 @@
+REACT_APP_AUTH0_DOMAIN=FILL-IT-IN
+REACT_APP_AUTH0_CLIENT_ID=FILL-IT-IN
+REACT_APP_AUTH0_AUDIENCE=FILL-IT-IN
