+++ javascript/.gitignore
@@ -13,12 +13,14 @@
 
 # misc
 .DS_Store
+!.env
 .env.local
 .env.development.local
 .env.test.local
+!.env.production.SAMPLE
+!.env.local.SAMPLE
 .env.production
 .env.production.local
-
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
