+++ pom.xml
@@ -8,10 +8,10 @@
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>edu.ucsb</groupId>
-	<artifactId>spring-react-minimal</artifactId>
+	<artifactId>mapache-search</artifactId>
 	<version>0.1.0</version>
-	<name>Spring React Minimal</name>
-	<description>Spring React Minimal (CMPSC 156 class project)</description>
+	<name>Mapache Search</name>
+	<description>Mapache Search (CMPSC 156 class project)</description>
 
 	<properties>
 		<java.version>11</java.version>
@@ -199,7 +199,7 @@
 					<archive>
 						<manifest>
 							<!-- full package name of class with the main you want to run -->
-							<mainClass>edu.ucsb.changeme.DemoApplication</mainClass>
+							<mainClass>edu.ucsb.mapache.DemoApplication</mainClass>
 
 						</manifest>
 					</archive>
