--- src/main/java/edu/ucsb/changeme/DemoApplication.java
@@ -1,17 +1,0 @@
-package edu.ucsb.changeme;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
-
-@Configuration
-@PropertySource(value = "classpath:secrets-localhost.properties", ignoreResourceNotFound = true)
-@SpringBootApplication
-public class DemoApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(DemoApplication.class, args);
-  }
-
-}
