--- src/main/java/edu/ucsb/changeme/controllers/ReactController.java
@@ -1,12 +1,0 @@
-package edu.ucsb.changeme.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class ReactController {
-  @RequestMapping(value = {"/", "/{x:[\\w\\-]+}", "/{x:^(?!api$).*$}/**/{y:[\\w\\-]+}"})
-  public String getIndex() {
-    return "/index.html";
-  }
-}
