--- src/main/java/edu/ucsb/changeme/repositories/AdminRepository.java
@@ -1,11 +1,0 @@
-package edu.ucsb.changeme.repositories;
-
-import edu.ucsb.changeme.entities.Admin;
-import java.util.List;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AdminRepository extends CrudRepository<Admin, Long> {
-  public List<Admin> findByEmail(String email);
-}
