--- src/main/java/edu/ucsb/changeme/repositories/AppUserRepository.java
@@ -1,11 +1,0 @@
-package edu.ucsb.changeme.repositories;
-
-import edu.ucsb.changeme.entities.AppUser;
-import java.util.List;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface AppUserRepository extends CrudRepository<AppUser, Long> {
-  public List<AppUser> findByEmail(String email);
-}
