+++ src/main/java/edu/ucsb/mapache/DemoApplication.java
@@ -1,0 +1,17 @@
+package edu.ucsb.mapache;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+
+@Configuration
+@PropertySource(value = "classpath:secrets-localhost.properties", ignoreResourceNotFound = true)
+@SpringBootApplication
+public class DemoApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(DemoApplication.class, args);
+  }
+
+}
