+++ src/main/java/edu/ucsb/mapache/controllers/ReactController.java
@@ -1,0 +1,12 @@
+package edu.ucsb.mapache.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+public class ReactController {
+  @RequestMapping(value = { "/", "/{x:[\\w\\-]+}", "/{x:^(?!api$).*$}/**/{y:[\\w\\-]+}" })
+  public String getIndex() {
+    return "/index.html";
+  }
+}
